moved chain migration to election types (#109)
