+++ planetmint/commands/planetmint.py
@@ -15,12 +15,12 @@ import json
 import sys
 
 from planetmint.core import rollback
-from planetmint.migrations.chain_migration_election import ChainMigrationElection
 from planetmint.utils import load_node_key
 from planetmint.transactions.common.transaction_mode_types import BROADCAST_TX_COMMIT
 from planetmint.transactions.common.exceptions import (
     DatabaseDoesNotExist, ValidationError)
 from planetmint.transactions.types.elections.vote import Vote
+from planetmint.transactions.types.elections.chain_migration_election import ChainMigrationElection
 import planetmint
 from planetmint import (backend, ValidatorElection,
                         Planetmint)
