+++ tests/elections/test_election.py
@@ -4,7 +4,7 @@ from tests.utils import generate_election, generate_validators
 
 from planetmint.lib import Block
 from planetmint.transactions.types.elections.election import Election
-from planetmint.migrations.chain_migration_election import ChainMigrationElection
+from planetmint.transactions.types.elections.chain_migration_election import ChainMigrationElection
 from planetmint.upsert_validator.validator_election import ValidatorElection
 
 @pytest.mark.bdb
